poc: make things super lazy
