+++ packages/manager/src/features/NodeBalancers/NodeBalancerDetail/NodeBalancerSummary/NodeBalancerSummary.tsx
@@ -1,14 +1,14 @@
 import { useNodeBalancerQuery } from '@linode/queries';
 import Grid from '@mui/material/Grid';
-import { useParams } from '@tanstack/react-router';
-import * as React from 'react';
+import { createLazyRoute, useParams } from '@tanstack/react-router';
+import React from 'react';
 
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 
 import { SummaryPanel } from './SummaryPanel';
 import { TablesPanel } from './TablesPanel';
 
-export const NodeBalancerSummary = () => {
+const NodeBalancerSummary = () => {
   const { id } = useParams({ from: '/nodebalancers/$id' });
   const { data: nodebalancer } = useNodeBalancerQuery(id);
 
@@ -26,3 +26,9 @@ export const NodeBalancerSummary = () => {
     </div>
   );
 };
+
+export const nodeBalancerSummaryLazyRoute = createLazyRoute(
+  '/nodebalancers/$id/summary'
+)({
+  component: NodeBalancerSummary,
+});
