+++ packages/manager/src/features/NodeBalancers/NodeBalancersLanding/NodeBalancerActionMenu.tsx
@@ -57,7 +57,7 @@ export const NodeBalancerActionMenu = (props: Props) => {
       onClick: () => {
         navigate({
           params: {
-            id: String(nodeBalancerId),
+            id: nodeBalancerId,
           },
           to: `/nodebalancers/$id/delete`,
         });
